Add `remove_object_store()`/`rename_object_store()` to `Databasebuilder` to make reuse useful (#41)
* build: add `indexmap` dependency

This allows to refer to `ObjectStoreBuilder`s by name and still keeping
the insertion order.

* chore: use `IndexMap` for object stores in `DatabaseBuilder`

This will make handling removals more straightforward.

* feat: add `DatabaseBuilder.remove_object_store()`

* feat: add `Databasebuilder.rename_object_store()`

* test: test added remove and rename methods

* feat: support combination of remove and rename

Internally, we cannot rely on just implicit removal when combining
remove and rename: if a store is removed and another store is then
renamed to the removed store name, implicit removal would not result in
the correct sequence of operations.

* test: test combination of remove and rename